Title bar and submit button update
