+++ client/src/components/CreateGames.js
@@ -15,6 +15,8 @@ import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
+import TitleBar from "./TitleBar";
+import SubmitButton from "./SubmitButton";
 
 
 
@@ -57,28 +59,7 @@ const styles = theme => ({
 
 });
 
-function ButtonAppBar(props) {
-  const { classes } = props;
-  return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton className={classes.menuButton} color="inherit" aria-label="Menu">
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" color="inherit" className={classes.grow}>
-            News
-          </Typography>
-          <Button color="inherit">Login</Button>
-        </Toolbar>
-      </AppBar>
-    </div>
-  );
-}
 
-ButtonAppBar.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
 
 class CreateGames extends React.Component {
   state = {
@@ -108,7 +89,7 @@ class CreateGames extends React.Component {
       <div style={{width:"100%"}}>
       <form className={classes.container} noValidate autoComplete="off">
         
-          
+          <TitleBar/>
           
           
           <TextField
@@ -201,7 +182,7 @@ class CreateGames extends React.Component {
           timeCaption="time"
       />
 
-  
+      <SubmitButton/>
 
         
         </form>
@@ -214,4 +195,4 @@ CreateGames.propTypes = {
   classes: PropTypes.object.isRequired
 };
 
-export default withStyles(styles)(CreateGames); withStyles(styles)(ButtonAppBar);
+export default withStyles(styles)(CreateGames);
